need to see what the repo returns as I make query dynamic

+++ backend/src/main/java/com/job_intel/backend/Dtos/HiringCompaniesDto.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class HiringCompaniesDto {
 
     String companyName;
-    Long jobGrowthRate;
+    Double jobGrowthRate;
     Long numOfJobs;
 
 }
